+++ trilby-cli/src/Trilby/Infect/Options.hs
@@ -21,7 +21,7 @@ data InfectOpts m = InfectOpts
 
 parseOpts :: forall m. (forall a. Parser a -> Parser (m a)) -> Parser (InfectOpts m)
 parseOpts f = do
-    hosts <- f . fmap (fromListSafe Localhost) . many $ fromString <$> strArgument (metavar "HOST" <> help "Target host to infect")
+    hosts <- f . parseHosts $ help "Hosts to infect, default: localhost"
     edition <- f $ parseEnum (long "edition" <> metavar "EDITION" <> help "The Trilby edition to infect with")
     reboot <- f $ parseYesNo "reboot" "Reboot when done infecting"
     authorisedKeys <- parseAuthorisedKeys f
@@ -40,7 +40,7 @@ parseAuthorisedKeys f = f do
 askOpts :: InfectOpts Maybe -> InfectOpts App
 askOpts opts =
     InfectOpts
-        { hosts = maybe (pure [Localhost]) pure opts.hosts
+        { hosts = askHosts opts.hosts
         , edition = maybe (selectEnum "Choose edition:" Nothing) pure opts.edition
         , authorisedKeys = maybe askAuthorisedKeys pure opts.authorisedKeys
         , reboot = maybe (yesNoButtons "Reboot system?" True) pure opts.reboot